% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_waterinfo.R
\name{call_waterinfo}
\alias{call_waterinfo}
\title{http call to waterinfo.be}
\usage{
call_waterinfo(query, base_url = "download", token = NULL)
}
\arguments{
\item{query}{list of query options to be used together with the base string}

\item{base_url}{str download | pro, default download defined}

\item{token}{token to use with the call (optional, can be retrieved via \code{\link{get_token}})}
}
\value{
waterinfo_api class object with content and info about call
}
\description{
General call used to request information and data from waterinfo.be,
providing error handling and json parsing
}
