% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_period.R
\name{check_date_format}
\alias{check_date_format}
\title{Check if the date can be parsed to a datetim object in R}
\usage{
check_date_format(datetime)
}
\arguments{
\item{datetime}{string representation of the date}
}
\value{
POSIXct date-time object is date is valid representation
}
\description{
if the date is already a datetime object ("POSIXct" "POSIXt"), the object
itself is returned
}
