% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_period.R
\name{check_period_format}
\alias{check_period_format}
\title{Check if the format of the period is conform the specifications of VMM}
\usage{
check_period_format(period_string)
}
\arguments{
\item{period_string}{input string according to format required by waterinfo:
The period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`,
each # is an integer value and the codes define the number of...
Y - years
M - months
D - days
T required if information about sub-day resolution is present
H - hours
D - days
M - minutes
S - seconds
Instead of D (days), the usage of W - weeks is possible as well
Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.}
}
\value{
str period string itself if valid
}
\description{
Check if the format of the period is conform the specifications of VMM
}
