% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_points.R
\name{get_raster_points}
\alias{get_raster_points}
\title{Get timeseries data for specific points on a raster}
\usage{
get_raster_points(
  ts_id = 911010,
  period = NULL,
  from = NULL,
  to = NULL,
  points,
  datasource = 10,
  token = NULL
)
}
\arguments{
\item{ts_id}{waterinfo.be database ts_id, defining a timeserie variable and
frequency it is defined.}

\item{period}{input string according to format required by waterinfo:
De period string is provided as P#Y#M#DT#H#M#S, with P defines `Period`,
each # is an integer value and the codes define the number of...
Y - years
M - months
D - days
T required if information about sub-day resolution is present
H - hours
D - days
M - minutes
S - seconds
Instead of D (days), the usage of W - weeks is possible as well
Examples of valid period strings: P3D, P1Y, P1DT12H, PT6H, P1Y6M3DT4H20M30S.}

\item{from}{date of datestring as start of the time series}

\item{to}{date of datestring as end of the time series}

\item{points}{a matrix or data.frame with coordinates in decimal degrees
using the EPSG 4326 coordinate system.
It must have the colnames `lat` and `long`.}

\item{datasource}{int [0-4] defines the `meetnet` of which the measurement
station is part of. VMM based stations are net '1', MOW-HIC is net '2'}

\item{token}{token to use with the call (optional, can be retrieved via
\code{\link{get_token}})}
}
\description{
Get timeseries data for specific points on a raster
}
