% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Get list of stations for a variable}
\usage{
get_stations(variable_name = NULL, frequency = "15min")
}
\arguments{
\item{variable_name}{char valid nam of available variable as timeseriesgroup}

\item{frequency}{char valid frequency for the given variable, for most
variables, the 15min frequency is available}
}
\description{
For a given timeseriesgroup (variable), provide a list of measurement
stations providing data. An overview of the variables is provided by the
function `supported_variables`
}
\details{
For the moment, this only works for measurement stations of VMM (meetnet 1),
and stations from other measurement nets are not included in the list
}
\seealso{
supported_variables
}
