% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_variables.R
\name{is_supported_variable}
\alias{is_supported_variable}
\title{Check if variable is supported by VMM ts group id}
\usage{
is_supported_variable(variable_name)
}
\arguments{
\item{variable_name}{char}
}
\value{
Raise error when variable is not supported directly, otherwise NULL
}
\description{
Check if variable is supported by VMM ts group id
}
\examples{
is_supported_variable("wind_speed")
}
