% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liedekerke}
\alias{liedekerke}
\title{Soil moisture data of Liedekerke, January 2017}
\format{A data frame with 23,816 rows and 9 variables:
\describe{
  \item{ts_id}{identifier of the downloaded time serie}
  \item{Timestamp}{datetime}
  \item{Value}{measured value of the variable}
  \item{Quality Code}{Quality code of the measurement}
  \item{station_name}{full name of the measurement station}
  \item{station_no}{short code name of the measurement station}
  \item{ts_name}{type/frequency of the time serie}
  \item{parametertype_name}{parameter type name}
  \item{stationparameter_name}{parameter name on station level}
}}
\source{
\url{https://www.waterinfo.be/}
}
\usage{
liedekerke
}
\description{
A dataset compiled by downloading 1 day of soil moisture data for the
Liedekerke  measurement station of Waterinfo.be
}
\keyword{datasets}
