% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_period.R
\name{parse_period}
\alias{parse_period}
\title{Handle the information of provided date information on the period and provide
feedback to the user}
\usage{
parse_period(from = NULL, to = NULL, period = NULL)
}
\arguments{
\item{from}{string representing date of datetime object}

\item{to}{string representing date of datetime object}

\item{period}{input string according to format required by waterinfo}
}
\value{
list with the relevant period/date information
}
\description{
valid combinations of the arguments are:
from/to, from/period, to/period, period, from
}
\seealso{
check_period_format
}
