% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_identifiers.R
\name{resolve_timeseriesgroupid}
\alias{resolve_timeseriesgroupid}
\title{Translate the usage of available variables to the corresponding
timeseriesgroupID, based on the provided lookup table from VMM}
\usage{
resolve_timeseriesgroupid(variable_name, frequency = "15min")
}
\arguments{
\item{variable_name}{valid variable name, supported by VMM API}

\item{frequency}{valid frequency for the given variable}
}
\description{
Remark that this information is NOT based on a query, but on information
provided by the package itself to make variable names more readable
}
\details{
The lookup table is provided as external data of the package,
see inst/extdata
}
