% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_identifiers.R
\name{resolve_timeseriesid}
\alias{resolve_timeseriesid}
\title{Translate the usage of measurement station identifiers in the user interface
and the required variable to the corresponding timeseriesID}
\usage{
resolve_timeseriesid(variable_name, station_no)
}
\description{
Translate the usage of measurement station identifiers in the user interface
and the required variable to the corresponding timeseriesID
}
