% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_variables.R
\name{supported_frequencies}
\alias{supported_frequencies}
\title{VMM supported timeseriesgroups frequencies}
\usage{
supported_frequencies(variable_name)
}
\arguments{
\item{variable_name}{char name of a valid variable in either dutch or english}
}
\description{
Provide list of VMM supported frequencies for a given timeseriesgroupID
in either dutch or english
}
\examples{
supported_frequencies('rainfall')
}
