% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_variables.R
\name{supported_variables}
\alias{supported_variables}
\title{VMM supported timeseriesgroups variables}
\usage{
supported_variables(language = "nl")
}
\arguments{
\item{language}{char nl (dutch) or en (english) variable names}
}
\description{
Provide list of VMM supported variables in the timeseriesgroupID
in either dutch or english
}
