% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_waterinfo.R
\name{call_waterinfo}
\alias{call_waterinfo}
\title{Basic http call to waterinfo.be, providing error handling and json parsing}
\usage{
call_waterinfo(query, base_url = "download")
}
\arguments{
\item{query}{list of query options to be used together with the base string}

\item{base_url}{str download | pro, default download defined}
}
\value{
waterinfo_api class object with content and info about call
}
\description{
Basic http call to waterinfo.be, providing error handling and json parsing
}
