% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_period.R
\name{isdatetime}
\alias{isdatetime}
\title{Check if the string input can be converted to a date, provides FALSE or date}
\usage{
isdatetime(datetime)
}
\arguments{
\item{datetime}{string representation of a date}
}
\value{
FALSE | "POSIXct" "POSIXt"
}
\description{
(acknowledgements to micstr/isdate.R)
}
\keyword{internal}
