% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_period.R
\name{isdatetime}
\alias{isdatetime}
\title{Check if the string input can be converted to a date}
\usage{
isdatetime(datetime)
}
\arguments{
\item{datetime}{string representation of a date}
}
\value{
boolean
}
\description{
(acknowledgments to micstr/isdate.R)
}
\examples{
isdatetime("1985-11-21")

}
