% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_waterinfo.R
\name{print.waterinfo_api}
\alias{print.waterinfo_api}
\title{Custom print function of the API request response}
\usage{
\method{print}{waterinfo_api}(x, ...)
}
\arguments{
\item{x}{waterinfo_api}
}
\description{
Custom print function of the API request response
}
