% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_identifiers.R
\name{resolve_datasource}
\alias{resolve_datasource}
\title{Define the datasource using the station number}
\usage{
resolve_datasource(station_no)
}
\arguments{
\item{station_no}{'stations-nummer' as it appears on the download page of
https://www.waterinfo.be/default.aspx?path=NL/Rapporten/Downloaden}
}
\description{
Using the 'stations-nummer' as provided on
https://www.waterinfo.be/default.aspx?path=NL/Rapporten/Downloaden, this
function tries to identify the datasource to use for the particular variable
}
\details{
Notice that VMM did not provide this in the official documentation, but this
has just been derived by checking the API response as such
}
\examples{
resolve_datasource('akl03e-1066')
resolve_datasource('K07_OM421')
}
