% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_variables.R
\name{supported_variables}
\alias{supported_variables}
\title{VMM supported timeseriesgroups variables}
\usage{
supported_variables(language = "nl")
}
\arguments{
\item{language}{char \code{nl} (dutch) or \code{en} (english) variable names}
}
\value{
data.frame containing the variable names in either english or dutch
}
\description{
Provide list of VMM supported variables in the timeseriesgroupID
in either dutch or english
}
\examples{
# Request supported variables in Dutch
supported_variables("nl")

# Request supported variables in English
supported_variables("en")
}
