% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRinfo.R
\docType{package}
\name{wateRinfo-package}
\alias{wateRinfo}
\alias{wateRinfo-package}
\title{wateRinfo: Download Time Series Data from Waterinfo.be}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The wateRinfo package provides an R interface to the data available
    at waterinfo.be, the data portal provided by the Flemish Environmental
    Agency. The package provides R functions to check the available stations for 
    a given variable, check the available variable for a given station and 
    download time series using the database identifiers as well as station 
    names.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/wateRinfo}
  \item \url{https://ropensci.github.io/wateRinfo}
  \item Report bugs at \url{https://github.com/ropensci/wateRinfo/issues}
}

}
\author{
\strong{Maintainer}: Stijn Van Hoey \email{stijn.vanhoey@inbo.be} (0000-0001-6413-3185)

Other contributors:
\itemize{
  \item Willem Maetens \email{w.maetens@vmm.be} [contributor]
  \item Peter Desmet \email{peter.desmet@inbo.be} (0000-0002-8442-8025) [contributor]
  \item Research Institute for Nature and Forest (INBO) \email{info@inbo.be} [copyright holder]
}

}
\keyword{internal}
