% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateRinfo.R
\docType{package}
\name{wateRinfo-package}
\alias{wateRinfo}
\alias{wateRinfo-package}
\title{wateRinfo: Download Time Series Data from Waterinfo.be}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

wateRinfo facilitates access to waterinfo.be 
    (<https://www.waterinfo.be>), a website managed by the Flanders Environment 
    Agency (VMM) and Flanders Hydraulics Research. The website provides access 
    to real-time water and weather related environmental variables for Flanders 
    (Belgium), such as rainfall, air pressure, discharge, and water level. The 
    package provides functions to search for stations and variables, and 
    download time series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/wateRinfo}
  \item \url{https://docs.ropensci.org/wateRinfo}
  \item Report bugs at \url{https://github.com/ropensci/wateRinfo/issues}
}

}
\author{
\strong{Maintainer}: Stijn Van Hoey \email{stijnvanhoey@gmail.com} (\href{https://orcid.org/0000-0001-6413-3185}{ORCID})

Other contributors:
\itemize{
  \item Willem Maetens \email{w.maetens@vmm.be} [contributor]
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) [contributor]
  \item Research Institute for Nature and Forest (INBO) \email{info@inbo.be} [copyright holder]
}

}
\keyword{internal}
